requirements.txt
git+https://github.com/Voxel-Fox-Ltd/Novus
databases
pillow
psutil
disnake-jishaku
PyNacl
git+https://github.com/6days9weeks/pomice
humanize
hentai
colorama
toml
voxelbotutils
tabulate
async-cse